+++ packages/codegen/src/templates/boilerplate/gitignore.ts
@@ -2,7 +2,6 @@ export const gitIgnoreTemplate = `
 dist
 coverage
 node_modules
-.eslintcache
 .env.local
 *.tsbuildinfo